+++ panda.config.ts
@@ -130,4 +130,6 @@ export default defineConfig({
 
   // The output directory for your css system
   outdir: "styles",
+  
+  jsxFramework: "react",
 });
