+++ src/components/Select.tsx
@@ -152,7 +152,15 @@ export function Select({
         data-selected={!!pickedOption?.value || undefined}
         {...props}
       >
-        <span className="label">{pickedOption?.label}</span>
+        <span
+          className="label"
+          // Initial value can depend on document
+          suppressHydrationWarning
+          // Will default to initial value no matter what if not dangerouslySetInnerHTML
+          dangerouslySetInnerHTML={{
+            __html: pickedOption?.label,
+          }}
+        />
         <CaretDown size={18} weight="light" className={caret} ref={caretRef} />
       </div>
       <div className={dropdown} ref={dropdownRef}>
