+++ src/types/components/Layout.d.ts
@@ -1,0 +1,1 @@
+type LayoutProps = React.PropsWithChildren;
